Drop Elasticsearch5Mapping and make Elasticsearch6Mapping the base implementation
